+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- AudioSourceDictionary uses AudioDictionary class.
+
 ### Added
 - DefaultAudioMixer
 - Dependency to TextMeshPro package